Initial commit of code rearrangement. Split clumpr into high-level packages clustering, histogram, and reconcile.
