--- empress/clumpr/.gitignore
@@ -1,13 +1,0 @@
-graphs/*
-TreeLifeData/*
-example*
-.vscode
-.DS_Store
-**.newick
-*.pyc
-htmlcov*/**
-**/*.png
-errorTrees
-*.swp
-out.txt
-
